MR clean up commit
